refactor : 오브젝트 및 함수 이름 변경
